split env extension method out
