+++ src/AzureAuth/AuthModeExtensions.cs
@@ -22,7 +22,7 @@ namespace Microsoft.Authentication.AzureAuth
         /// <returns>A single aggregate <see cref="AuthMode"/> filtered for interactivity.</returns>
         public static AuthMode CombinedAuthMode(this IEnumerable<AuthMode> authModes, IEnv env)
         {
-            if (InteractiveAuthDisabled(env))
+            if (env.InteractiveAuthDisabled())
             {
 #if PlatformWindows
                 return AuthMode.IWA;
@@ -32,17 +32,6 @@ namespace Microsoft.Authentication.AzureAuth
             }
 
             return authModes.Aggregate((a1, a2) => a1 | a2);
-        }
-
-        /// <summary>
-        /// Determines whether interactive auth is allowed or not.
-        /// </summary>
-        /// <param name="env">An <see cref="IEnv"/> to use.</param>
-        /// <returns>A boolean to indicate PCA is disabled.</returns>
-        public static bool InteractiveAuthDisabled(IEnv env)
-        {
-            return !string.IsNullOrEmpty(env.Get(EnvVars.NoUser)) ||
-                string.Equals("1", env.Get(EnvVars.CorextNonInteractive));
         }
     }
 }
